Move utils/stream.js
