+++ utils/stream.js
@@ -1,4 +1,4 @@
-import { config } from "./config/index.js"
+import { config } from "../config/index.js"
 import { Client } from "twitter-api-sdk"
 
 export const stream = () => {
